added twitter.html in calendar.html
In this way it was easy to place the calendar table next to the twitter feeds as we had designed in Miro. 

I don't particularly like it this layout but let's have a discussion all together when we present it? 

Also, I have left Haris's feed as the csc twitter does not have tweets. We will only need to replace the link in the html in case we ever decide to put the csc twitter isntead.