Set up C4 page ready for fellows, moved application page to new C5 page
Made the fellow page URLs and picture URLs dynamically generated on the layout pages, instead of static links from the YML file which are time-consuming to generate per fellow in future cohorts.
Several photos changed to JPG from PNG to point to correct paths.
Removed or replaced a couple of dead links in fellow profiles